Fix memberships POST handler to return correct collection
Previously, the handler attempted to retrieve a collection using
the ID returned from inserting into the collections_images table,
which caused an IndexError since the ID does not refer to a
collection.

This commit changes the query to use the existing collection_id
from the request payload, ensuring the correct collection is returned.